\name{loadedPackages}
\Rdversion{1.1}
\alias{loadedPackages}
\title{The list of loaded packages}
\description{Simple utility that returns the list of loaded packages}
\usage{
loadedPackages()
}
\value{
A character vector containing the list of loaded packages
}
\author{
Romain Francois <francoisromain@free.fr>
}
\keyword{ manip }
