\name{svTools-package}
\alias{svTools-package}
\alias{svTools}
\docType{package}

\title{ SciViews GUI API - Tools (wrapper for packages tools and codetools) }
\description{
  Set of tools aimed at wrapping some of the functionalities of the packages
  tools, utils and codetools into a nicer format so that an IDE can use them.
}
\details{
  \tabular{ll}{
    Package: \tab svTools\cr
    Type: \tab Package\cr
    Version: \tab 0.9-1\cr
    Date: \tab 2012-03-31\cr
    License: \tab GPL 2 or above, at your convenience\cr
  }
  % TODO: add description of main functions here. Also add examples
}

\author{
  Romain Francois

  Maintainer: Ph. Grosjean <phgrosjean@sciviews.org>
}

\keyword{ package }
\keyword{ utilities }

\concept{ GUI API code tools and utilities for IDE }
